+++ app/src/main/java/com/forrestguice/suntimeswidget/SuntimesConfigActivity0.java
@@ -96,7 +96,7 @@ public class SuntimesConfigActivity0 extends AppCompatActivity
     private WidgetThemes.ThemeListAdapter spinner_themeAdapter;
     protected Spinner spinner_theme;
 
-    protected Spinner spinner_1x1mode;
+    protected Spinner spinner_1x1mode, spinner_3x2mode;
     protected CheckBox checkbox_allowResize;
     protected CheckBox checkbox_showTitle;
     protected TextView label_titleText;
@@ -497,10 +497,13 @@ public class SuntimesConfigActivity0 extends AppCompatActivity
         }
 
         //
-        // widget: 1x1 widget mode
+        // widget: 1x1 widget mode, 3x2 widget mode
         //
         spinner_1x1mode = (Spinner) findViewById(R.id.appwidget_appearance_1x1mode);
         initWidgetMode1x1(context);
+
+        spinner_3x2mode = (Spinner) findViewById(R.id.appwidget_appearance_3x2mode);
+        initWidgetMode3x2(context);
 
         //
         // widget: title text
@@ -680,6 +683,30 @@ public class SuntimesConfigActivity0 extends AppCompatActivity
     {
         WidgetSettings.WidgetModeSun1x1 mode1x1 = WidgetSettings.loadSun1x1ModePref(context, appWidgetId);
         spinner_1x1mode.setSelection(mode1x1.ordinal());
+    }
+
+    /**
+     * @param context a context used to access resources
+     */
+    protected void initWidgetMode3x2(Context context)
+    {
+        // EMPTY
+    }
+
+    /**
+     * @param context a context used to access resources
+     */
+    protected void saveWidgetMode3x2(Context context)
+    {
+        // EMPTY
+    }
+
+    /**
+     * @param context a context used to access resources
+     */
+    protected void loadWidgetMode3x2(Context context)
+    {
+        // EMPTY
     }
 
     /**
@@ -883,8 +910,9 @@ public class SuntimesConfigActivity0 extends AppCompatActivity
      */
     protected void saveAppearanceSettings(Context context)
     {
-        // save: widgetmode_1x1
+        // save: widgetmode_1x1, 3x2
         saveWidgetMode1x1(context);
+        saveWidgetMode3x2(context);
 
         // save: theme
         ThemeDescriptor theme = (ThemeDescriptor)spinner_theme.getSelectedItem();
@@ -915,8 +943,9 @@ public class SuntimesConfigActivity0 extends AppCompatActivity
      */
     protected void loadAppearanceSettings(Context context)
     {
-        // load: widgetmode_1x1
+        // load: widgetmode_1x1, 3x2
         loadWidgetMode1x1(context);
+        loadWidgetMode3x2(context);
 
         // load: theme
         SuntimesTheme theme = WidgetSettings.loadThemePref(context, appWidgetId);
@@ -1388,6 +1417,19 @@ public class SuntimesConfigActivity0 extends AppCompatActivity
         if (layout_1x1mode != null)
         {
             layout_1x1mode.setVisibility(View.GONE);
+        }
+    }
+
+    /**
+     *
+     * @param show true show option, false hide option (default hidden)
+     */
+    protected void showOption3x2LayoutMode(boolean show)
+    {
+        View layout_1x1mode = findViewById(R.id.appwidget_appearance_3x2mode_layout);
+        if (layout_1x1mode != null)
+        {
+            layout_1x1mode.setVisibility(show ? View.VISIBLE : View.GONE);
         }
     }