+++ app/src/main/java/com/forrestguice/suntimeswidget/themes/SuntimesTheme.java
@@ -98,6 +98,11 @@ public class SuntimesTheme
     public static final String THEME_FALLCOLOR = "fallcolor";
     public static final String THEME_WINTERCOLOR = "wintercolor";
 
+    public static final String THEME_MAP_BACKGROUNDCOLOR = "mapbackgroundcolor";
+    public static final String THEME_MAP_FOREGROUNDCOLOR = "mapforegroundcolor";
+    public static final String THEME_MAP_SHADOWCOLOR = "mapshadowcolor";
+    public static final String THEME_MAP_HIGHLIGHTCOLOR = "maphighlightcolor";
+
     public static final String THEME_TITLESIZE = "titlesize";
     public static final float THEME_TITLESIZE_MIN = 6.0f;
     public static final float THEME_TITLESIZE_DEF = 10.0f;
@@ -165,6 +170,11 @@ public class SuntimesTheme
     protected int themeSummerColor;
     protected int themeFallColor;
     protected int themeWinterColor;
+
+    protected int themeMapBackgroundColor;
+    protected int themeMapForegroundColor;
+    protected int themeMapShadowColor;
+    protected int themeMapHighlightColor;
 
     protected int themeMoonriseTextColor;
     protected int themeMoonsetTextColor;
@@ -245,6 +255,11 @@ public class SuntimesTheme
         this.themeSummerColor = otherTheme.themeSummerColor;
         this.themeFallColor = otherTheme.themeFallColor;
         this.themeWinterColor = otherTheme.themeWinterColor;
+
+        this.themeMapBackgroundColor = otherTheme.themeMapBackgroundColor;
+        this.themeMapForegroundColor = otherTheme.themeMapForegroundColor;
+        this.themeMapShadowColor = otherTheme.themeMapShadowColor;
+        this.themeMapHighlightColor = otherTheme.themeMapHighlightColor;
 
         this.themeTitleSize = otherTheme.themeTitleSize;
         this.themeTextSize = otherTheme.themeTextSize;
@@ -334,6 +349,11 @@ public class SuntimesTheme
         this.themeFallColor = themes.getInt( theme + THEME_FALLCOLOR, defaultTheme.themeFallColor );
         this.themeWinterColor = themes.getInt( theme + THEME_WINTERCOLOR, defaultTheme.themeWinterColor );
 
+        this.themeMapBackgroundColor = themes.getInt(theme + THEME_MAP_BACKGROUNDCOLOR, defaultTheme.themeMapBackgroundColor);
+        this.themeMapForegroundColor = themes.getInt(theme + THEME_MAP_FOREGROUNDCOLOR, defaultTheme.themeMapForegroundColor);
+        this.themeMapShadowColor = themes.getInt(theme + THEME_MAP_SHADOWCOLOR, defaultTheme.themeMapShadowColor);
+        this.themeMapHighlightColor = themes.getInt(theme + THEME_MAP_HIGHLIGHTCOLOR, defaultTheme.themeMapHighlightColor);
+
         this.themeTitleSize = themes.getFloat( theme + THEME_TITLESIZE, defaultTheme.themeTitleSize );
         this.themeTextSize = themes.getFloat( theme + THEME_TEXTSIZE, defaultTheme.themeTextSize );
         this.themeTimeSize = themes.getFloat( theme + THEME_TIMESIZE, defaultTheme.themeTimeSize );
@@ -410,6 +430,11 @@ public class SuntimesTheme
         themePrefs.putInt(themePrefix + SuntimesTheme.THEME_SUMMERCOLOR, this.themeSummerColor);
         themePrefs.putInt(themePrefix + SuntimesTheme.THEME_FALLCOLOR, this.themeFallColor);
         themePrefs.putInt(themePrefix + SuntimesTheme.THEME_WINTERCOLOR, this.themeWinterColor);
+
+        themePrefs.putInt(themePrefix + SuntimesTheme.THEME_MAP_BACKGROUNDCOLOR, this.themeMapBackgroundColor);
+        themePrefs.putInt(themePrefix + SuntimesTheme.THEME_MAP_FOREGROUNDCOLOR, this.themeMapForegroundColor);
+        themePrefs.putInt(themePrefix + SuntimesTheme.THEME_MAP_SHADOWCOLOR, this.themeMapShadowColor);
+        themePrefs.putInt(themePrefix + SuntimesTheme.THEME_MAP_HIGHLIGHTCOLOR, this.themeMapHighlightColor);
 
         themePrefs.putFloat(themePrefix + SuntimesTheme.THEME_TITLESIZE, this.themeTitleSize);
         themePrefs.putFloat(themePrefix + SuntimesTheme.THEME_TEXTSIZE, this.themeTextSize);
@@ -493,6 +518,9 @@ public class SuntimesTheme
         themePrefs.remove(themePrefix + SuntimesTheme.THEME_SUMMERCOLOR);
         themePrefs.remove(themePrefix + SuntimesTheme.THEME_FALLCOLOR);
         themePrefs.remove(themePrefix + SuntimesTheme.THEME_WINTERCOLOR);
+
+        themePrefs.remove(themePrefix + SuntimesTheme.THEME_MAP_BACKGROUNDCOLOR);
+        themePrefs.remove(themePrefix + SuntimesTheme.THEME_MAP_FOREGROUNDCOLOR);
 
         themePrefs.remove(themePrefix + SuntimesTheme.THEME_TITLESIZE);
         themePrefs.remove(themePrefix + SuntimesTheme.THEME_TEXTSIZE);
@@ -764,6 +792,26 @@ public class SuntimesTheme
             default:
                 return getSpringColor();
         }
+    }
+
+    public int getMapBackgroundColor()
+    {
+        return themeMapBackgroundColor;
+    }
+
+    public int getMapForegroundColor()
+    {
+        return themeMapForegroundColor;
+    }
+
+    public int getMapShadowColor()
+    {
+        return themeMapShadowColor;
+    }
+
+    public int getMapHighlightColor()
+    {
+        return themeMapHighlightColor;
     }
 
     public float getTitleSizeSp()