+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * adds permissions READ_CALENDAR, WRITE_CALENDAR; needed to interact w/ Calendar app (add/remove events in custom calendars).
 * adds permissions READ_SYNC_STATS, WRITE_SYNC_SETTINGS; needed to provide custom calendars (add/remove calendars via SyncAdapter).
 * adds a SyncAdapter (LOCAL account) to provide the Calendar app with custom calendars.
+* adds 3x1 and 3x2 SunPosition previews to theme editor.
+* adds 3x2 SunPosition widget showing world map.
+* adds world map dialog to app; shows sunlight (day/night) and moonlight over an equirectangular map.
 * adds theme "Dark (translucent)"; default theme with semitransparent widget background.
 * adds to theming; custom background option (simple background color supporting transparency).